fix MLX lib; apply changes from helper;;
Also add raw vesrion of button poller.

+++ mxl90640_lib/MLX90640_API.cpp
@@ -858,7 +858,7 @@ void ExtractPTATParameters(uint16_t *eeData, paramsMLX90640 *mlx90640)
 
 void ExtractGainParameters(uint16_t *eeData, paramsMLX90640 *mlx90640)
 {
-    int16_t gainEE;
+    int32_t gainEE;
     
     gainEE = eeData[48];
     if(gainEE > 32767)
@@ -1055,7 +1055,7 @@ void ExtractOffsetParameters(uint16_t *eeData, paramsMLX90640 *mlx90640)
     int occRow[24];
     int occColumn[32];
     int p = 0;
-    int16_t offsetRef;
+    int32_t offsetRef;
     uint8_t occRowScale;
     uint8_t occColumnScale;
     uint8_t occRemScale;
@@ -1125,11 +1125,11 @@ void ExtractOffsetParameters(uint16_t *eeData, paramsMLX90640 *mlx90640)
 void ExtractKtaPixelParameters(uint16_t *eeData, paramsMLX90640 *mlx90640)
 {
     int p = 0;
-    int8_t KtaRC[4];
-    int8_t KtaRoCo;
-    int8_t KtaRoCe;
-    int8_t KtaReCo;
-    int8_t KtaReCe;
+    int16_t KtaRC[4];
+    int16_t KtaRoCo;
+    int16_t KtaRoCe;
+    int16_t KtaReCo;
+    int16_t KtaReCe;
     uint8_t ktaScale1;
     uint8_t ktaScale2;
     uint8_t split;
